requirements.txt
wheel
flask
flask-wtf
flask-sqlalchemy
flask-bcrypt
flask-login
Pillow
flask-mail
